Remove extraneous service base
